Move DataBlobReference to its own module
Split the storage models code so that each model has their own python
module.
