+++ django/thunderstore/storage/models/__init__.py
@@ -1,1 +1,2 @@
 from .blob import *
+from .reference import *
